# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: Drug_Presciption_App
  description: An application that allows a user to keep track of their drug prescriptions
  contact: {}
host: drug-prescription-app.herokuapp.com
basePath: /api/v1
schemes:
- https
consumes:
- application/json
produces:
- application/json
paths:
...
  /formula:
    get:
      summary: Get usage formula
      tags:
      - usage formula
      operationId: Getusageformula
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: _id
        in: query
        required: true
        type: integer
        format: int64
        description: ''
      - name: token
        in: header
        required: true
        type: string
        description: ''
      responses:
        200:
          description: ''
          headers: {}
...